Margin fixes and h2 style change
Margins should be 15px uniformed through main and from header to main and main to footer. h2 no longer has a box shadow, while the different color still pulls your eyes to it, it fades in more than the important h1. The body no longer has calculated/set heights, relying on margin to make the content in between main and footer to grow or shrink depending on screen and content size.